Seems all functions work properly from Develop (#8)
* add is_in_shoping_cart field into RecipeSerializer

* add ability to mark recipes as favorite

* fix SubscriptionsSerializer response

* fix users/me permission

* add SEARCH_PARAM = 'name'

* add slug field into tag serializer

* add AauthorOrReadOnly permission

* add duplicate ingredients validation

* additional validation when create recipes

* fix shopping_cart and favorited fields

* add recipes_count field in SubscriptionsSerializer

* add recipe exist validation when favorite recipes

* fix patchmethod  for partial_update action

* add get_object_or_validation_error, recipes_limit

* add AllValuesMultipleFilter in RecipeFilter

* fix update method for RecipeCreateSerializer

* add FavoritedRecipeFilter

* add is_in_shopping_cart recipes filter

* crope unsupported filter_backends

* fix SubscriberSerializer post and delete actions

* fix APIRecipeCard delete and post actions

* fix FavoriteRecipeSerializer: post, delete actions

---------

+++ backend/api/views.py
@@ -1,62 +1,96 @@
 from django.shortcuts import get_object_or_404
 
-from django_filters import CharFilter
+from django_filters import AllValuesMultipleFilter
 from django_filters import rest_framework as filter
-from rest_framework import permissions, status
+from rest_framework import filters, mixins, permissions, status, viewsets
+from rest_framework.decorators import action
 from rest_framework.pagination import LimitOffsetPagination
 from rest_framework.response import Response
 from rest_framework.views import APIView
-from rest_framework.viewsets import ModelViewSet, ReadOnlyModelViewSet
 
+from api.permissions import AauthorOrReadOnly
 from api.serializers import IngredientSerializer, TagSerializer
-from recipes.models import Ingredient, Recipe, ShoppingCart, Tag
-from recipes.serializers import (RecipeCreateSerializer, RecipeSerializer,
+from api.utilities import get_object_or_validation_error
+from recipes.models import (FavoriteRecipe, Ingredient, Recipe, ShoppingCart,
+                            Tag)
+from recipes.serializers import (FavoriteRecipeSerializer,
+                                 RecipeCreateSerializer, RecipeSerializer,
+                                 RecipeSubscriptionSerializer,
                                  ShoppingCartSerializer)
 
 
-class TagViewSet(ReadOnlyModelViewSet):
+class TagViewSet(viewsets.ReadOnlyModelViewSet):
     queryset = Tag.objects.all()
     permission_classes = [permissions.IsAuthenticatedOrReadOnly]
     serializer_class = TagSerializer
     pagination_class = None
 
 
-class IngredientViewSet(ReadOnlyModelViewSet):
+class IngredientViewSet(viewsets.ReadOnlyModelViewSet):
     queryset = Ingredient.objects.all()
     permission_classes = [permissions.IsAuthenticatedOrReadOnly]
     serializer_class = IngredientSerializer
     pagination_class = None
+    filter_backends = [filters.SearchFilter]
+    search_fields = ['^name']
 
 
-class RecipeFilter(filter.FilterSet):
-    tags = CharFilter(field_name='tags__slug')
-    ingredient = CharFilter(field_name='ingredients__name')
+class FavoriteRecipeViewSet(mixins.CreateModelMixin, mixins.DestroyModelMixin,
+                            viewsets.GenericViewSet):
+    http_method_names = 'post', 'delete'
+    permission_classes = [permissions.IsAuthenticated]
+    queryset = FavoriteRecipe.objects.all()
+    serializer_class = FavoriteRecipeSerializer
+    pagination_class = None
+
+    @action(detail=False, methods=['delete'])
+    def delete(self, request, id):
+        recipe = get_object_or_404(Recipe, pk=id)
+        delete_record = FavoriteRecipe.objects.filter(
+            user=request.user, recipe=recipe)
+        if len(delete_record):
+            delete_record.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response(data={
+            "error_recipe": 'Такой записи нет в избранном.'},
+            status=status.HTTP_400_BAD_REQUEST)
+
+
+class RecipeFieldsFilter(filter.FilterSet):
+    tags = AllValuesMultipleFilter(field_name='tags__slug')
+    ingredient = AllValuesMultipleFilter(field_name='ingredients__name')
 
     class Meta:
         model = Recipe
-        fields = (
-            'author', 'tags', 'cooking_time', 'ingredient'
-        )
+        fields = ('author', 'tags', 'cooking_time', 'ingredient')
+
+    def filter_queryset(self, queryset):
+        if self.request.user.is_anonymous:
+            return super().filter_queryset(queryset)
+        if self.request.query_params.get('is_favorited', '0') == '1':
+            queryset = queryset.filter(favored__user=self.request.user)
+        if self.request.query_params.get('is_in_shopping_cart', '0') == '1':
+            queryset = queryset.filter(shoppers__user=self.request.user)
+        return super().filter_queryset(queryset)
 
 
-class RecipeViewSet(ModelViewSet):
+class RecipeViewSet(viewsets.ModelViewSet):
     permission_classes = [permissions.IsAuthenticatedOrReadOnly]
     queryset = Recipe.objects.all()
     pagination_class = LimitOffsetPagination
-    filterset_class = RecipeFilter
+    filterset_class = RecipeFieldsFilter
 
     def perform_create(self, serializer):
         serializer.save(author=self.request.user)
 
     def get_serializer_class(self):
-        if self.action in ('create',):
+        if self.action in ('create', 'detail', 'partial_update'):
             return RecipeCreateSerializer
-
         return RecipeSerializer
 
     def get_permissions(self):
         if self.action in ['destroy', 'partial_update']:
-            permission_classes = [permissions.IsAdminUser]
+            permission_classes = [AauthorOrReadOnly]
         elif self.action in ['create', 'retrieve', 'list']:
             permission_classes = [permissions.IsAuthenticatedOrReadOnly]
         else:
@@ -70,14 +104,16 @@ class APIRecipeCard(APIView):
     MEASURE = 'ingredient__ingredient__measurement_unit'
     AMOUNT = 'ingredient__amount'
 
+    http_method_names = ['get', 'post', 'delete']
+
     def get(self, request):
         if request._user.is_authenticated:
             ingredients_list = Recipe.objects.values(
                 self.ID, self.NAME, self.MEASURE, self.AMOUNT).filter(
                 shoppers__user=request.user
             )
-            shop_list = self.summ_amount(ingredients_list)
-            return Response(shop_list, status=status.HTTP_200_OK)
+            shopping_cart = self.summ_amount(ingredients_list)
+            return Response(shopping_cart, status=status.HTTP_200_OK)
         return Response(status=status.HTTP_401_UNAUTHORIZED)
 
     def summ_amount(self, obj_dict):
@@ -87,23 +123,24 @@ class APIRecipeCard(APIView):
                 summ_list[obj[self.ID]] = obj
             else:
                 summ_list[obj[self.ID]][self.AMOUNT] += obj[self.AMOUNT]
-        shop_cart = [
-            ''.join((row[self.NAME], ': ',
-                     str(row[self.AMOUNT]), row[self.MEASURE]))
-            for row in summ_list.values()
-        ]
-        return shop_cart
+
+        cart = [{row[self.NAME]: (row[self.AMOUNT], row[self.MEASURE])}
+                for row in summ_list.values()]
+        return {'Список необходимых ингредиентов': cart}
 
     def post(self, request, id):
         if request._user.is_anonymous:
             return Response(status=status.HTTP_401_UNAUTHORIZED)
-        recipe = get_object_or_404(Recipe, pk=id)
+        recipe = get_object_or_validation_error(
+            Recipe, id, 'Неверный ID рецепта.')
         request.data['user'] = request.user.id
         request.data['shop_it'] = recipe.id
         serializer = ShoppingCartSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
+            recipe_serializer = RecipeSubscriptionSerializer(recipe)
+            return Response(
+                recipe_serializer.data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, id):
@@ -112,5 +149,9 @@ class APIRecipeCard(APIView):
         recipe = get_object_or_404(Recipe, pk=id)
         delete_record = ShoppingCart.objects.filter(
             user=request.user, shop_it=recipe)
-        delete_record.delete()
-        return Response(None, status.HTTP_204_NO_CONTENT)
+        if len(delete_record):
+            delete_record.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response(data={
+            "error_recipe": 'Такой записи в корзине нет.'},
+            status=status.HTTP_400_BAD_REQUEST)
