+++ backend/users/views.py
@@ -15,6 +15,11 @@ class CustomUserViewSet(UserViewSet):
     pagination_class = LimitOffsetPagination
     permission_classes = [permissions.IsAuthenticatedOrReadOnly]
 
+    def get_permissions(self):
+        if self.action == "me":
+            return [permissions.IsAuthenticated()]
+        return super().get_permissions()
+
 
 class SubscriberViewSet(ModelViewSet):
     permission_classes = [permissions.IsAuthenticated]
@@ -30,8 +35,12 @@ class SubscriberViewSet(ModelViewSet):
         author = get_object_or_404(User, pk=author_id)
         delete_record = Subscriber.objects.filter(
             subscribe=request.user, author=author)
-        delete_record.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if len(delete_record):
+            delete_record.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response(data={
+            "error_id": 'Такой подписки не существует'},
+            status=status.HTTP_400_BAD_REQUEST)
 
 
 class SubscriptionsViewSet(ModelViewSet):
@@ -39,6 +48,7 @@ class SubscriptionsViewSet(ModelViewSet):
     http_method_names = 'get',
     serializer_class = SubscriptionsSerializer
     queryset = User.objects.all()
+    pagination_class = LimitOffsetPagination
 
     def get_queryset(self):
         return self.request.user.subscription.all()
